TagsPlugin: Make realm configuration for `ListTagged` more effective, refs #7857 and #9063.
Settings for 'listtagged_exclude_realms' had no effect without a query.
Itamar Ostricher discovered this issue in the initial implementation of the
feature and suggested this change.  Thanks for taking care.

Additionally the broken 'test_empty_content' test in
`tractags.tests.macros.ListTaggedMacroTestCase` is fixed, and a regression test
for the pending 'empty query' issue (#7857) has been introduced as well.

git-svn-id: https://trac-hacks.org/svn/tagsplugin/trunk@12391 7322e99d-02ea-0310-aa39-e9a107903beb