TagsPlugin: Mirror ticket tags into normalized `tags` db table, refs #4503, #9194 and #11096.
Parsing ticket fields per request has been a known performance hog, that's why
using the `tags` db table for regular access yields a considerable speed-up.

As a side-effect 'ticket' is the pilot for fixing a recently announced issue
with `get_tagged_resources` in the default tag provider implementation.
Nicely polished by stretching unit test coverage to this part of the code.

'''ToDo:''' `DefaultTagProvider` must be fixed as well, and other tag
providers should be reviewed too as a precaution.

git-svn-id: https://trac-hacks.org/svn/tagsplugin/trunk@13165 7322e99d-02ea-0310-aa39-e9a107903beb